requirements.txt
elevenlabs
google-genai
sounddevice
numpy
pyserial
matplotlib
scipy
